pyspark/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.3
numpy==1.22.3
pandas==1.4.2
pkg_resources==0.0.0
protobuf==3.20.0
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
venv-pack==0.2.0
